Add Current Clipboard Content as Entry menu option
Created a command in the context menu which will add whatever is in
the user's clipboard as a new entry, assuming it isn't whitespace or
already isnide of the entries. Useful if corresponding entry is deleted
and user doesn't want to past and re-copy a value to add it to the
entries list.